requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.1.2
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distlib==0.4.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.0
filelock==3.19.1
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
h11==0.16.0
httptools==0.7.1
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
pipenv==2025.0.4
platformdirs==4.3.8
plotly==6.0.1
polars==1.29.0
prompt_toolkit==3.0.52
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.0
pyarrow==20.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.30.0
uvloop==0.22.1
virtualenv==20.34.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==16.0
